homework7

+++ src/controllers/contacts.js
@@ -71,19 +71,21 @@ export const getContactByIdController = async (req, res) => {
 export const createContactController = async (req, res) => {
   let photo = null;
 
-  if (getEnvVar("UPLOAD_TO_CLOUDINARY") === "true") {
-    const result = await uploadToCloudinary(req.file.path);
+  if (req.file?.path) {
+    if (getEnvVar("UPLOAD_TO_CLOUDINARY") === "true") {
+      const result = await uploadToCloudinary(req.file.path);
 
-    await fs.unlink(req.file.path);
+      await fs.unlink(req.file.path);
 
-    photo = result.secure_url;
-  } else {
-    await fs.rename(
-      req.file.path,
-      path.resolve("src", "uploads", "photos", req.file.filename)
-    );
+      photo = result.secure_url;
+    } else {
+      await fs.rename(
+        req.file.path,
+        path.resolve("src", "uploads", "photos", req.file.filename)
+      );
 
-    photo = `http://localhost:3000/photos/${req.file.filename}`;
+      photo = `http://localhost:3000/photos/${req.file.filename}`;
+    }
   }
 
   const contact = await createContact({
